Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main(){
-    double a;
-    scanf("%lf",&a);
+    int a;
+    scanf("%d",&a);
     if(a&1){
         printf("Not Set");
     }
